requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.2
aniso8601==9.0.1
appdirs==1.4.4
argcomplete==1.8.1
async-timeout==4.0.3
attrs==19.1.0
Automat==20.2.0
Babel==2.8.0
bcrypt==4.2.0
bidict==0.23.1
blinker==1.9.0
build==1.2.2.post1
buildozer==1.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloud-init==24.3.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
Cython==3.0.11
dbus-python==1.2.18
Deprecated==1.2.14
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.21.2
exceptiongroup==1.2.2
Faker==26.0.0
filelock==3.15.4
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.7
importlib-metadata==4.6.4
importlib_resources==6.4.0
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
Kivy==2.3.0
Kivy-Garden==0.1.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
limits==3.13.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
maxminddb==2.6.2
mdurl==0.1.2
more-itertools==8.10.0
multidict==6.0.5
netifaces==0.11.0
npm==0.1.1
numpy==2.1.0
oauthlib==3.2.0
opencv-python==4.10.0.84
optional-django==0.1.0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
passlib==1.7.4
pexpect==4.9.0
pipenv==2024.0.1
platformdirs==4.2.2
proxybroker==0.3.2
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycares==4.5.0
pycparser==2.22
pycurl==7.44.1
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyproject_hooks==1.2.0
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-nmap==0.7.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==5.4.1
redis==5.0.8
requests==2.32.3
rich==13.7.1
scapy==2.5.0
sdkmanager==0.5.1
SecretStorage==3.3.1
selenium==4.24.0
service-identity==18.1.0
sh==1.14.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
sqlalchemy-serializer==1.4.22
systemd-python==234
texttable==1.6.4
toml==0.10.2
tomli==2.2.1
trio==0.26.2
trio-websocket==0.11.1
twilio==9.2.3
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.2
virtualenv==20.26.3
virtualenv-clone==0.3.0
wadllib==1.3.6
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==1.0.0
zope.interface==5.4.0
